chore: action to ensure dev PRs only have 1 commit
We cannot configure GitHub to require different PR
merging strategies for each branch

To be safe to land PR's on main which needs to keep commits
the same as they are in dev we've configured to only allow rebase
merging

At the same time we want all changes for PRs to dev to be
squashed into a single commit. Therefore, this action will
enforce that CI checks only pass if a PR has a single commit.
When there are multiple commits, the PR will need to be squashed
manually/forced pushed to one before it is landed.

Signed-off-by: Michael Dawson <midawson@redhat.com>